third_party/WebKit/LayoutTests/external/wpt/html/browsers/history/the-history-interface/traverse_the_history_4-expected.txt

This is a testharness.js-based test.
FAIL Multiple history traversals, last would be aborted assert_array_equals: Pages opened during history navigation property 1, expected 5 but got 3
Harness: the test ran to completion.
